requirements.txt
affine==2.4.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contextily==1.3.0
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
Fiona==1.9.1
fonttools==4.38.0
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
mercantile==1.2.1
munch==2.5.0
numpy==1.24.2
-e git+ssh://git@github.com/openskynetwork/opensky-api.git@0a4259acee4e17d7e9b2f88e1d3c59d52f01afbf#egg=opensky_api&subdirectory=python
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pymongo==4.3.3
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
rasterio==1.3.6
requests==2.28.2
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
urllib3==1.26.14
xyzservices==2023.2.0

setuptools~=65.5.1
pip~=22.3.1
wheel~=0.38.4
Brotli~=1.0.9
Jinja2~=3.1.2
MarkupSafe~=2.1.2
Flask~=2.2.3
APScheduler~=3.10.0